+++ sshd-common/src/test/java/org/apache/sshd/common/config/keys/loader/openssh/OpenSSHKeyPairResourceParserDecodingTest.java
@@ -18,24 +18,18 @@
  */
 package org.apache.sshd.common.config.keys.loader.openssh;
 
-import java.net.URL;
 import java.security.KeyPair;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.util.Collection;
 import java.util.List;
 
-import org.apache.sshd.common.config.keys.AuthorizedKeyEntry;
 import org.apache.sshd.common.config.keys.BuiltinIdentities;
-import org.apache.sshd.common.config.keys.FilePasswordProvider;
 import org.apache.sshd.common.config.keys.KeyUtils;
 import org.apache.sshd.common.config.keys.PrivateKeyEntryDecoder;
-import org.apache.sshd.common.config.keys.PublicKeyEntryResolver;
 import org.apache.sshd.common.util.GenericUtils;
 import org.apache.sshd.util.test.JUnit4ClassRunnerWithParametersFactory;
-import org.apache.sshd.util.test.JUnitTestSupport;
 import org.apache.sshd.util.test.NoIoTestCase;
-import org.junit.Assume;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -52,16 +46,10 @@ import org.junit.runners.Parameterized.UseParametersRunnerFactory;
 @RunWith(Parameterized.class)   // see https://github.com/junit-team/junit/wiki/Parameterized-tests
 @UseParametersRunnerFactory(JUnit4ClassRunnerWithParametersFactory.class)
 @Category({ NoIoTestCase.class })
-public class OpenSSHKeyPairResourceParserTest extends JUnitTestSupport {
-    private static final OpenSSHKeyPairResourceParser PARSER = OpenSSHKeyPairResourceParser.INSTANCE;
-    private static final String PASSWORD = "super secret passphrase";
-    private static final FilePasswordProvider PASSWORD_PROVIDER = FilePasswordProvider.of(PASSWORD);
-    private static final String ENCRYPTED_RESOURCE_PREFIX = "encrypted";
+public class OpenSSHKeyPairResourceParserDecodingTest extends OpenSSHKeyPairResourceParserTestSupport {
 
-    private final BuiltinIdentities identity;
-
-    public OpenSSHKeyPairResourceParserTest(BuiltinIdentities identity) {
-        this.identity = identity;
+    public OpenSSHKeyPairResourceParserDecodingTest(BuiltinIdentities identity) {
+        super(identity);
     }
 
     @Parameters(name = "type={0}")
@@ -71,41 +59,19 @@ public class OpenSSHKeyPairResourceParserTest extends JUnitTestSupport {
 
     @Test
     public void testLoadUnencryptedKeyPairs() throws Exception {
-        testLoadKeyPairs(false);
+        testLoadKeyPairs(false, null);
     }
 
     @Test
     public void testLoadEncryptedKeyPairs() throws Exception {
-        testLoadKeyPairs(true);
+        testLoadKeyPairs(true, DEFAULT_PASSWORD_PROVIDER);
     }
 
-    private void testLoadKeyPairs(boolean encrypted) throws Exception {
-        Assume.assumeTrue(identity + " not supported", identity.isSupported());
-
-        String resourceKey = getClass().getSimpleName() + "-" + identity.getName().toUpperCase() + "-" + KeyPair.class.getSimpleName();
-        if (encrypted) {
-            resourceKey = ENCRYPTED_RESOURCE_PREFIX + "-" + resourceKey;
-        }
-
-        URL urlKeyPair = getClass().getResource(resourceKey);
-        if (encrypted) {
-            Assume.assumeTrue(identity + " no encrypted test data", urlKeyPair != null);
-        } else {
-            assertNotNull("Missing key-pair resource: " + resourceKey, urlKeyPair);
-        }
-
-        Collection<KeyPair> pairs = PARSER.loadKeyPairs(null, urlKeyPair, PASSWORD_PROVIDER);
+    @Override
+    protected void testLoadKeyPairs(
+            boolean encrypted, String resourceKey, Collection<KeyPair> pairs, PublicKey pubEntry)
+                throws Exception {
         assertEquals("Mismatched pairs count", 1, GenericUtils.size(pairs));
-
-        URL urlPubKey = getClass().getResource(resourceKey + ".pub");
-        assertNotNull("Missing public key resource: " + resourceKey, urlPubKey);
-
-        List<AuthorizedKeyEntry> entries = AuthorizedKeyEntry.readAuthorizedKeys(urlPubKey);
-        assertEquals("Mismatched public keys count", 1, GenericUtils.size(entries));
-
-        AuthorizedKeyEntry entry = entries.get(0);
-        PublicKey pubEntry = entry.resolvePublicKey(null, PublicKeyEntryResolver.FAILING);
-        assertNotNull("Cannot retrieve public key", pubEntry);
 
         Class<? extends PublicKey> pubType = identity.getPublicKeyType();
         Class<? extends PrivateKey> prvType = identity.getPrivateKeyType();